chore: small changes
1. removed chinese comments
2. fixed bash script error

+++ build.sh
@@ -1,7 +1,7 @@
 cd react
 npm install
 npm run build
-rm -r ../public/markdown-viewer
+[ -d ../public/markdown-viewer ] && rm -r ../public/markdown-viewer
 mkdir -p ../public/markdown-viewer
 cp -r dist/* ../public/markdown-viewer/
